🔢 Combine array utils into single file
Also fixes a mis-named joinList test file
