+++ src/utils/index.ts
@@ -1,7 +1,7 @@
+export * from './array';
 export * from './fs';
 export * from './path';
 export * from './fileExtension';
-export { default as flatten } from './flatten';
 export { default as isTruthy } from './isTruthy';
 export { default as joinList } from './joinList';
 export { default as noop } from './noop';
